[Fix] make things prettier
