+++ jira/jira_tools/__init__.py
@@ -1,2 +1,1 @@
 from .tools import *
-from .tools_python import *
