+++ jira/jira_tools/tools/__init__.py
@@ -1,3 +1,1 @@
 from .issues import *
-from .epics import *
-from .sprints import *
