+++ jira/jira_tools/tools/basic_funcs.py
@@ -60,8 +60,3 @@ def get_jira_basic_headers() -> dict:
         "Accept": "application/json",
         "Content-Type": "application/json",
     }
-
-
-def get_file_content(path) -> bytes:
-    with open(path, "rb") as f:
-        return f.read()
