+++ jira/jira_tools/tools/list_issues.py
@@ -7,7 +7,14 @@ from basic_funcs import (
 import requests
 
 
-def list_issues_in_project(project_key, num_issues=5, status=None, assignee=None, priority=None, reporter=None):
+def list_issues_in_project(
+    project_key: str,
+    num_issues: int = 5,
+    status: str = None,
+    assignee: str = None,
+    priority: str = None,
+    reporter: str = None,
+):
     jql_query = f"project = {project_key}"
     if status:
         jql_query += f" AND status = '{status}'"
@@ -23,9 +30,14 @@ def list_issues_in_project(project_key, num_issues=5, status=None, assignee=None
         "maxResults": num_issues,
         "fields": "summary,created",
     }
+
     try:
-        search_url = f"{ATLASSIAN_JIRA_API_URL}/{get_jira_cloud_id()}/rest/api/3/search/jql"
-        response = requests.get(search_url, headers=get_jira_basic_headers(), params=params)
+        search_url = (
+            f"{ATLASSIAN_JIRA_API_URL}/{get_jira_cloud_id()}/rest/api/3/search/jql"
+        )
+        response = requests.get(
+            search_url, headers=get_jira_basic_headers(), params=params
+        )
         response.raise_for_status()
 
         issues = response.json().get("issues", [])
@@ -38,6 +50,7 @@ def list_issues_in_project(project_key, num_issues=5, status=None, assignee=None
             }
             for issue in issues
         ]
+
         return latest_issues
 
     except requests.exceptions.RequestException as e:
@@ -50,31 +63,37 @@ def main():
 
     parser = argparse.ArgumentParser(description="List Jira issues")
     parser.add_argument("project_key", help="Jira Project key")
-    parser.add_argument("--issues_number", default=5,type=int, help="Number of issue to list")
-    parser.add_argument("--status", default=None,type=str, help="Issues status, such as Done")
-    parser.add_argument("--assignee", default=None,type=str, help="including assignee user")
-    parser.add_argument("--priority", default=None,type=str, help="including issues priority")
-    parser.add_argument("--reporter", default=None,type=str, help="including assignee reporter")
+    parser.add_argument(
+        "--issues_number", default=5, type=int, help="Number of issue to list"
+    )
+    parser.add_argument(
+        "--status", default=None, type=str, help="Issues status, such as Done"
+    )
+    parser.add_argument(
+        "--assignee", default=None, type=str, help="including assignee user"
+    )
+    parser.add_argument(
+        "--priority", default=None, type=str, help="including issues priority"
+    )
+    parser.add_argument(
+        "--reporter", default=None, type=str, help="including assignee reporter"
+    )
     args = parser.parse_args()
 
-    no_value ='<no value>'
-    issues_number = args.issues_number if not no_value else 5
-    status = args.status if not no_value else None
-    assignee = args.assignee if not no_value else None
-    priority = args.priority if not no_value else None
-    reporter = args.reporter if not no_value else None
-
+    no_value = "<no value>"  # when no value is injected
     try:
         latest_issues = list_issues_in_project(
             args.project_key,
-            issues_number,
-            status,
-            assignee,
-            priority,
-            reporter
+            args.issues_number if not no_value else 5,
+            args.status if not no_value else None,
+            args.assignee if not no_value else None,
+            args.priority if not no_value else None,
+            args.reporter if not no_value else None,
         )
         for issue in latest_issues:
-            print(f"{issue['key']}: {issue['summary']} - Created on {issue['created']})")
+            print(
+                f"{issue['key']}: {issue['summary']} - Created on {issue['created']})"
+            )
     except Exception as e:
         print(f"Failed to create issue: {e}")
