--- jira/jira_tools/tools_python/__init__.py
@@ -1,1 +1,0 @@
-from .issues import *
